maximal 2-edge-connected components (#14)
