+++ benches/common/factories.rs
@@ -1,6 +1,9 @@
+use std::fs::File;
+use std::io::Read;
 use std::marker::PhantomData;
 
 use petgraph::data::Build;
+use petgraph::graph6::FromGraph6;
 use petgraph::prelude::*;
 use petgraph::visit::NodeIndexable;
 
@@ -330,4 +333,13 @@ pub fn directed_fan(n: usize) -> DiGraph<(), ()> {
     }
 
     g
+}
+
+/// Parse a file in graph6 format into an undirected graph
+pub fn ungraph_from_graph6_file(path: &str) -> Graph<(), (), Undirected, u32> {
+    let mut f = File::open(path).expect("file not found");
+    let mut contents = String::new();
+    f.read_to_string(&mut contents)
+        .expect("failed to read from file");
+    Graph::from_graph6_string(contents)
 }
